Fixing comments vol2
