+++ src/components/Home/Home.scss
@@ -1,0 +1,8 @@
+.subtitle{
+    text-align: left;
+    margin-bottom: 10px;
+    h2 {
+        margin-bottom: 0px;
+    }
+
+}
